requirements.txt
psycopg2-binary==2.7.6
pytest==4.0.0
requests==2.20.1
